Enable edge back gesture on app launch
We don't clear state on EdgeBackGestureHandler
when we go to Recents or Home.
A more robust solution is in the works with
WM team and will probably be built on top of this.
Needs ag/11273364 to work.

Bug: 154580671
Test: Launch app directly in landscape,
should be able to swipe home on first try.

Change-Id: Ia4c685e16645d0237f278a611f6c61f252f3b79c
(cherry picked from commit 3f3cc3e96e453edea3b77a3909063c5da590412a)