Notify SysUi on device rotation for back gesture
Whenever device rotates, we notify sysui to hide/show
the back gesture if the foreground app also rotates.
We also notify sysui when device rotates to match
the orientation of the current foreground app (this
is for apps with fixed rotations).

Fixes: 154580671
Test: Created test apps of different rotations
and ensured that back functionality was present when
attempting to go back.

Change-Id: I33a71698411d9bc2416b6660f8dbd53233628917