Maintain fake nav bar position in overview for landscape
NavBar defaults to display when quickswitch ends,
which is always portrait since home is now always
in portrait. We keep the same nav region that the
user used to swipe up to enter recents and reset once
they swipe to home.

Fixes: 156053957
Test: Swipe up from landscape app, then swipe up
from the same nav bar region to go home.

Change-Id: Ia4cf8172850b991aa5e145897622be3b2afa3779