Start quickswitch on task switch instead of touch down
Previously we would enter a quickswitch session state
whenever user tapped on navbar and sent flags to
sysui to only enable edge gesture for the current rotation.
Now we only do that on a new task resulting from the gesture
or when going into recents to show nav bar in the
orientation user would expect it (only if they never
quickswitched).

Fixes: 155844780
Test: Tested swiping back from quickswitching
into apps of different orientations.
Note that back is still broken when you
quickswitch and then rotate the foreground app.

Change-Id: I935e7c66489d46e59156414d64653c846ae4c5d8