Update TaskView fullscreen params when updating the thumbnail matrix
Previously, we were only updating the fullscreen params when updating
fullscreen progress. The problem with that is we optimize to skip
updating fullscreen progress if the new progress is the same as the
current. But since the fullscreen params depend on the insets computed
by the preview position helper, we should update the fullscreen params
when we update the preview position helper.

Test: force stop launcher, quick switch directly to the right, ensure
the adjacent TaskView has its insets applied properly

Bug: 157105795
Change-Id: I3c3b47c69205607c9e10ed38c403211816f30939