Set mLastRectRotation to valid rotation
mLastRectRotation's value is used after
swipe gestures are complete to reset the
layout for overview (after swiping to home
animation is complete), so we need to set
it to a valid value instead of -1.
OverviewActions visibility depends on
it being rotation 0 for it to be visible,
otherwise it won't show up.

Bug: 155844780
Test: Swipe up to home from an app,
Swipe up to overview, observe overview
actions show in portrait.
Swipe up and back down from landscape app,
hit back. Open portrait app, ensure back
gesture still works.

Change-Id: Ia923a54e68af567b4470841783328528fffebe6a