Fix bug where scrim did not fade in during staggered workspace animation.
The problem is that in prepareToAnimate, we set the scrim sysUiProgress to 1.

When we call setScrim(Background) we create an animator to set sysUiProgress
to 0 with a duration of 0. We expect this to set sysUiProgress to 0 prior
to the next animation setScrirm(Normal).

When we call setScrim(Normal) we want to animate sysUiProgress to 1, but it
is still 1 since we haven't actually run the setScrim(Background) animation
(we've only added it to the list of animators). Therefore we don't create an
animator at all since the target is already at the final value.

Bug: 158148281
Change-Id: I8ca7ef00fc8a7dad6813c285ead60f5db15939b6