[Overview Actions] For in-place landscape, add select only when aiai works.
When display orientation is not the same as device orientation, aiai won't work for select mode highlight. So we should only show the select button in task menu for in-place landscape when the display rotation is the same as the device rotation which is portrait.

Screenshot would be in the task menu for in-place landscape all the time.

Test: demo video: https://drive.google.com/file/d/1OQrjww2lFTTsthz4dB3BSa-jw943KLWp/view?usp=sharing
Change-Id: I58b6f6da234ea36eef0af5bb966c33ebda3625b8