Don't reset GestureState when touching outside of swipe region
We were setting mGestureState = DEFAULT_STATE when touching outside
of the swipe region. Instead, we should only set mGestureState when
starting a new gesture, and only set it to DEFAULT_STATE on reset.

This also makes it clearer that mGestureState always refers to the
current gesture state, rather than sometimes being the current
gesture and sometimes being the previous gesture depending on
where it was used. Now previousGestureState is passed when needed.

Bug: 159494933
Change-Id: Ib62278b7c591ca4c6b99b26e058a34772e6cec2a