Increasing wait time for end-action events.
This is supposed to fix a flake in one of widgets tests.

The end-scroll event is posted by the system server to another thread
may arrive in ~13 sec.

The delay may have been caused by the previous test that just created
and deleted a user, so the system is busy with processing that.

Bug: 160238801
Change-Id: I43d0804252202ae04c731f35fb219c4be4bd4a76
(cherry picked from commit ba161b7393994481b84581525be7f76331d7d15b)