Fix hotseat and prediction row to allow updates when empty.
Rotating the screen in the homescreen empties the hotseat, however it does not get populated while it is visible to the user. The user should not be able to see an empty hotseat or prediction row if predictions are available. It should therefore be possible to populate these when they are empty even if they are visible to the user.

Change-Id: I8e5252bd29050c2cd9d443aedcb3f3e305c0e2d7