Fixing animation cancelled not called if gesture was not initiated
Previously, in case of a tap in the gesture region, swipeHandler was not
explicitly cancelling the launcher state animation. It was getting reset
bucause launcher gets onStop immediately after that.
But in some cases, it may not get onStop (if the top activity finishes),
which leaves the Launcher UI in a broken state

Bug: 172842997
Change-Id: Iff7b1e5b8f67b2f4dd8096b88c4cd97115ec912a