Removing blocking call during cuj_LAUNCHER_APP_LAUNCH_FROM_ICON
This is a part of Blue Chip de-janking effort and is
a P0 item in our CUJ list.

In the last frame of the app-open animation, we have
a
android.view.IRemoteAnimationFinishedCallback$Stub$Proxy.onAnimationFinished call
which is blocking and prone to lock contention in WindowManager.

Moving the call to after the animation.

Bug: 173072373
Change-Id: If5eb6469900d1892b861eda6d6b4807d90630b6b