Resolve Flaky work tab tests
In cases where test execution and launcher relayout overlap, mLauncher.getAppsView().getContentView() returns a recyclerview instead of pagedView. This is resolved when rebindAdapters is later called with showTabs=true. The issue here is we are keeping a stale value of mLauncher.getAppsView().getContentView() in WorkEduView which is problematic as pagedView is required to set the right tab index for the test.

This behavior also affects work testWorkEduIntermittent as we need to verify the a pagedView is visible before text execution.
Bug: 149867607
Bug: 159671700
Test: flake -t com.android.launcher3.ui.WorkTabTest

Change-Id: I4b0968d6b9daa5c8fe741d4b1ed090f9c7acabed