Added new app launch crop / positioning.
New app launch crops the app window from the center, which is meant
to align with the centered icon in the new preview window, and then
it "reveals" the rest of the app window from center to out.

Code is behind same prop as the shell starting surface.

To enable:
- adb shell set prop persist.debug.shell_starting_surface true
- reboot

Bug: 177264697
Change-Id: I78f2eab216d4bb40dedae245f47a89fe677dd6e9