[Live Tile] Fix a bunch of issues related to touch input
Here are the issues
- Touch down again during transition of quick switch brings back the previous app
- During transition to overview, touch down on task / in the background area results in weird behaviors

These issues related to touch input are caused by the z-ordering of app and launcher. Because in the current implementation we place launcher on top upon initial swipe up, during the transition (to overview / quick switch), any touch event goes directly to launcher instead of the overview input consumer (where it normally goes to in the non-live tile case), resulting in these weird behaviors mentioned above.

The reason we put app below launcher upon swipe up is to render task view icon earlier (before swipe up settles into overview), which, is not as critical (we don't have the feature now), and we can find other ways to make that happen.

So here are the proposed changes
- Place app on top upon swipe up. Only place launcher on top after transition to overview.
- Remove the temporary fix to disable click for live tile task during transition.
- Remove icon logic in live tile overlay (in this change I only removed the call, but let me know if it's preferred that they are completely removed from LiveTileOverlay in this change)

Test: manual
Fixes: 178640174
Fixes: 175039524
Change-Id: Ib97804c8bbc487bc55160719375c0f2ff3ba7fd3