Continue receiving key events from input consumer after settling into Overview
This builds on ag/13732971. In live tile mode after settling into overview, recents_animation_input_consumer should have the focus, and thus we will need to extend the lifecycle of input consumers so that key events can be routed to launcher

Overall, since live tile extends the lifecycle of certain private variables of AbsSwipeHandler, I am thinking of doing a refactor by putting those things into a separate class, but it might be too big for this change.

TODO: I've verified that this change works while if we don't process motion events, it would cause b/182932863. Will investigate on that.

Fixes: 181404912
Test: Cherry pick ag/13732971; Go to overview; Swipe back; Observe the animation to go back to app
Change-Id: I5afd397be387b75f373442781dd1d00560dca99e