Consolidate two different onTransitionCancelled implementations
There was previously two different onTransitionCancelled(), one in
BaseActivityInterface and one in AnimationFactory. They are pretty
similar and the differences seem subtle and not necessary. In this
bug, the AnimationFactory version was going to mStartState instead
of StateManager#getRestState() (which is the same, unless mStartState
disables restore). Instead, always use the mActivityInterface
implemenation, which seems more robust.

Test: Open app A, open app B, quick switch from B to A, swipe home
during that transition, then touch nav region during home transtion;
ensure we don't goToState(BACKGROUND_APP) in the subsequent
onTransitionCanceled().

Fixes: 184583037

Change-Id: I315a5087b12413d88c3ccc1e3245a55a76a73e62