Iterate over all swipeable regions when calculating quickswitch gesture bounds
Previously when creating a new CurrentDisplay, we would
never get the correct OrientationRectF from the hashmap
because the size field of CurrentDisplay would be incorrect
for a given rotation.
Ex. Rotation 0 and Rotation 1 would have inverted display sizes
Thus hashcode would return different values and we would only
ever see if a point is contained in the current rotation's
OrientationRectF.

Bug: 183897242
Test: Tested swiping from portrait to landscape app, was
able to swipe continuously

Change-Id: I6de625389956f55cfc2142dcb7aeef2e90c157ac