Animate App Widget activity launch from Quickstep launchers
Following the general pattern on icon launch animations, define a
'floating widget view' to represent the appearance of the app widget
through the launch animation.

The floating widget view separates the foreground and background
components of the LauncherAppWidgetHostView, which can be positioned
and animated separately.

The background (or placeholder if no background can be identified)
moves and grows from its original position and size to the launched
app's position and size.

The widget's foreground scales and moves to match, fading out to be
replaced with the app window.

Bug: 169042867
Test: manual
Change-Id: I65d2b1bc80275f9df460790720e13d1650093347