Do not detect system gesture with new system ui state flag
If the device is in 2-button or gesturl navgiation mode,
touch events will be cancelled in the nav bar area. It leads
to that users couldn't drag the window magnification via
draghandle.

To fix it, we add a new system ui state flag to stop detecting
system gesture if it is set to true. This flag should be set only
when the window magnification is overlapped with the nav bar from
windowinsets.

Bug: 179648683
Test: manually test
Change-Id: I978a4b7d631e4115b293d341a89baee151c2881a